+++ cosmos/models/mobility/src/test/java/es/tid/cosmos/mobility/mapreduce/ParseCellsMapperTest.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.util.List;
 
 import com.twitter.elephantbird.mapreduce.io.ProtobufWritable;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mrunit.mapreduce.MapDriver;
@@ -21,19 +20,19 @@ import es.tid.cosmos.mobility.mapreduce.ParseCellsMapper;
  * @author dmicol
  */
 public class ParseCellsMapperTest {
-    private MapDriver<IntWritable, Text, LongWritable, ProtobufWritable<Cdr>>
+    private MapDriver<LongWritable, Text, LongWritable, ProtobufWritable<Cdr>>
             driver;
     
     @Before
     public void setUp() {
-        this.driver = new MapDriver<IntWritable, Text, LongWritable,
+        this.driver = new MapDriver<LongWritable, Text, LongWritable,
                 ProtobufWritable<Cdr>>(new ParseCellsMapper());
     }
 
     @Test
     public void test() throws IOException {
         List<Pair<LongWritable, ProtobufWritable<Cdr>>> results = this.driver
-                .withInput(new IntWritable(1),
+                .withInput(new LongWritable(1L),
                            new Text("33F43052|2221436242|12|34|56|78"))
                 .run();
         assertEquals(1, results.size());
